Twitter Updated filter
The new version will check if the URL submitted is `x.com`. and convertes it to twitter.com format. This converstion is executed as in the backed the data is still stored as `twitter.com`.

The future of twitter is currently not clear, as it might sold out after the US election and the new buyers might bring back the old name.

Fixes #82